Fix global rot y stability and Unreal files
The derivative of the global rotation on the 2d plane was unstable. This
has been fixed by making the values of rot y alwyas positive. When
taking the derivative one should be mindful to get the smallest angle.
Unreal files were missing from the repo, wrong (empty) files were in
their places.